+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - GeoTrellisPath assumes `file` scheme when none provided [#3191](https://github.com/locationtech/geotrellis/pull/3191)
 - toStrings overrides to common classes [#3217](https://github.com/locationtech/geotrellis/pull/3217)
+- GeoTrellisRasterSources legacy and temporal layers support [#3179](https://github.com/locationtech/geotrellis/issues/3179)
 
 ### Changed